requirements.txt
streamlit
requests
moviepy
google-generativeai
pydub
spacy
python-dotenv
assemblyai
ffmpeg-python
